% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_functions.R
\name{controlVariatesME}
\alias{controlVariatesME}
\title{Control variates method}
\usage{
controlVariatesME(
  df,
  X,
  Y,
  A,
  Astar,
  val_idx,
  sl.lib,
  estimand = "ATE",
  probs = NA
)
}
\arguments{
\item{df}{A dataframe containing the outcome, treatment and covariates}

\item{X}{A vector of covariates}

\item{Y}{A vector of outcomes}

\item{A}{A vector of treatments}

\item{Astar}{A vector of error-prone treatment measurements}

\item{val_idx}{A vector of validation data indicators}

\item{sl.lib}{A vector of super learner libraries to use}

\item{estimand}{The estimand of interest (e.g. ATE, ATT, ATC)}

\item{probs}{A vector of probabilities for sampling into val data}
}
\value{
A list containing the ATE est and associated variance
}
\description{
Control variates method
}
