% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalizability-functions.R
\name{controlVariatesMEGen}
\alias{controlVariatesMEGen}
\title{Main function for implementing the "generalizability" version of the control
variates estimator}
\usage{
controlVariatesMEGen(
  data,
  X,
  Y,
  A,
  Astar,
  val_idx,
  sl.lib,
  estimand = "ATE",
  rho = NA
)
}
\arguments{
\item{data}{A dataframe containing the outcome, treatment and covariates}

\item{X}{A vector of covariate names}

\item{Y}{String containing outcome name}

\item{A}{String containing treatment name}

\item{val_idx}{String containing validation index name}

\item{sl.lib}{A vector of super learner libraries to use}

\item{estimand}{The estimand of interest (either 'ATE' or 'ATT')}

\item{rho}{A vector of participation probabilities}
}
\value{
A list containing the ATE estimate, corresponding variance estimate,
and estimates of Gamma and V from the asymptotic covariance matrix
}
\description{
Main function for implementing the "generalizability" version of the control
variates estimator
}
