% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hajek_functions.R
\name{controlVariatesMEHajek}
\alias{controlVariatesMEHajek}
\title{Control variates method}
\usage{
controlVariatesMEHajek(
  df,
  X,
  Y,
  A,
  Astar,
  val_idx,
  probs,
  sl.lib,
  estimand = "ATE",
  Ystar = NA
)
}
\arguments{
\item{df}{A dataframe containing the outcome, treatment and covariates}

\item{X}{A vector of covariate names}

\item{Y}{A string containing the outcome name}

\item{A}{A string containing the treatment name}

\item{Astar}{A vector of error-prone treatment measurement names}

\item{val_idx}{A vector of validation data indicators}

\item{probs}{A vector of probabilities for sampling into val data}

\item{sl.lib}{A vector of super learner libraries to use}

\item{estimand}{The estimand of interest (e.g. ATE, ATT, ATC)}

\item{Ystar}{A vector of error-prone outcome measurement names}
}
\value{
A list containing the ATE est and associated variance
}
\description{
Control variates method
}
