% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalizability-functions.R
\name{generalizability_eif}
\alias{generalizability_eif}
\title{Function for obtaining the generalizability efficient influence function
terms}
\usage{
generalizability_eif(
  mu_hat1,
  mu_hat0,
  pi_hat,
  kappa_hat,
  data,
  X,
  Y,
  A,
  val_idx
)
}
\arguments{
\item{mu_hat1}{A vector of predicted outcomes under treatment}

\item{mu_hat0}{A vector of predicted outcomes under no treatment}

\item{pi_hat}{A vector of predicted propensity scores}

\item{kappa_hat}{A vector of predicted selection probabilities}

\item{data}{A dataframe containing the outcome, treatment and covariates}

\item{X}{A vector of covariates}

\item{Y}{A vector of outcomes}

\item{A}{A vector of treatments}

\item{val_idx}{A vector of validation indices}
}
\value{
A vector containing the estimated EIF
}
\description{
Function for obtaining the generalizability efficient influence function
terms
}
