% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalizability-functions.R
\name{generalizability_est}
\alias{generalizability_est}
\title{Function for implementing generalizability estimator}
\usage{
generalizability_est(data, X, Y, A, val_idx, sl.lib, rho = NA)
}
\arguments{
\item{data}{A dataframe containing the outcome, treatment and covariates}

\item{X}{A vector of covariates}

\item{Y}{A vector of outcomes}

\item{A}{A vector of treatments}

\item{val_idx}{A vector of validation indices}

\item{sl.lib}{A vector of super learner libraries to use}
}
\value{
A vector containing the estimated EIF
}
\description{
Function for implementing generalizability estimator
}
