% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_functions.R
\name{get_tau_hat}
\alias{get_tau_hat}
\title{Get val. estimate}
\usage{
get_tau_hat(df, X, Y, A, estimand, sl.lib)
}
\arguments{
\item{df}{A dataframe containing the outcome, treatment and covariates}

\item{X}{A vector of covariates}

\item{Y}{A vector of outcomes}

\item{A}{A vector of treatments}

\item{estimand}{The estimand of interest (e.g. ATE, ATT, ATC)}

\item{sl.lib}{A vector of super learner libraries to use}
}
\value{
A list containing the AIPW estimate of the estimand
}
\description{
Get val. estimate
}
