% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hajek_functions.R
\name{get_tau_hat_hajek}
\alias{get_tau_hat_hajek}
\title{Implement Hajék estimator -- rescale EIF of val. only estimator by function
of inverse sampling probs}
\usage{
get_tau_hat_hajek(df, X, Y, A, probs, sl.lib)
}
\arguments{
\item{df}{A dataframe containing the outcome, treatment and covariates}

\item{X}{A vector of covariates}

\item{Y}{A vector of outcomes}

\item{A}{A vector of treatments}

\item{sl.lib}{A vector of super learner libraries to use}

\item{estimand}{The estimand of interest (e.g. ATE, ATT, ATC)}
}
\value{
A list containing the AIPW estimate of the estimand
}
\description{
Implement Hajék estimator -- rescale EIF of val. only estimator by function
of inverse sampling probs
}
