% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_functions.R
\name{get_vcov_asym}
\alias{get_vcov_asym}
\title{Function for getting asymptotic variance of control variates estimator}
\usage{
get_vcov_asym(tau_val_mod, cv_mods, val_idx)
}
\arguments{
\item{tau_val_mod}{A list containing the AIPW estimate of the estimand}

\item{cv_mods}{A list containing the AIPW estimate of the control variates}

\item{val_idx}{A vector of validation data indicators}
}
\value{
A list containing estimates of V and Gamma
}
\description{
Function for getting asymptotic variance of control variates estimator
}
