% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalizability-functions.R
\name{is_binary}
\alias{is_binary}
\title{Function for checking if a vector is binary}
\usage{
is_binary(x)
}
\arguments{
\item{x}{A vector of values}
}
\value{
A boolean indicating if the vector is binary
}
\description{
Function for checking if a vector is binary
}
