% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalizability-functions.R
\name{kappa_hat_est}
\alias{kappa_hat_est}
\title{Function for estimating the probability of selection into the validation set}
\usage{
kappa_hat_est(data, X, val_idx, sl.lib)
}
\arguments{
\item{data}{A dataframe containing the outcome, treatment and covariates}

\item{X}{A vector of covariates}

\item{val_idx}{A vector of validation indices}

\item{sl.lib}{A vector of super learner libraries to use}
}
\value{
A vector containing the estimated probability of selection into the validation set
}
\description{
Function for estimating the probability of selection into the validation set
}
