% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalizability-functions.R
\name{pi_hat_est}
\alias{pi_hat_est}
\title{Function for estimating the propensity score model}
\usage{
pi_hat_est(data, X, A, val_idx, sl.lib)
}
\arguments{
\item{data}{A dataframe containing the outcome, treatment and covariates}

\item{X}{A vector of covariates}

\item{A}{A vector of treatments}

\item{sl.lib}{A vector of super learner libraries to use}
}
\value{
A vector containing the estimated propensity score model
}
\description{
Function for estimating the propensity score model
}
