% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalizability-functions.R
\name{mu_hat_est}
\alias{mu_hat_est}
\title{Function for estimating the outcome model}
\usage{
mu_hat_est(data, X, Y, A, val_idx, sl.lib)
}
\arguments{
\item{data}{A dataframe containing the outcome, treatment and covariates}

\item{X}{A vector of covariate names}

\item{Y}{String containing outcome name}

\item{A}{String containing treatment name}

\item{val_idx}{String containing validation index name}

\item{sl.lib}{A vector of super learner libraries to use}
}
\value{
A list containing the estimated outcome model
}
\description{
Function for estimating the outcome model
}
